chore: rename signal-utils to internal-api
